+++ crates/block-producer/src/mempool/mod.rs
@@ -53,6 +53,7 @@ use subscription::SubscriptionProvider;
 use tokio::sync::{Mutex, MutexGuard, mpsc};
 use tracing::{instrument, warn};
 
+use crate::domain::batch::SelectedBatch;
 use crate::domain::transaction::AuthenticatedTransaction;
 use crate::errors::{AddTransactionError, VerifyTxError};
 use crate::mempool::budget::BudgetStatus;
@@ -263,7 +264,7 @@ impl Mempool {
     ///
     /// Returns `None` if no transactions are available.
     #[instrument(target = COMPONENT, name = "mempool.select_batch", skip_all)]
-    pub fn select_batch(&mut self) -> Option<(BatchId, Vec<Arc<AuthenticatedTransaction>>)> {
+    pub fn select_batch(&mut self) -> Option<SelectedBatch> {
         // The selection algorithm is fairly neanderthal in nature.
         //
         // We iterate over all transaction nodes, each time selecting the first transaction which
@@ -278,13 +279,13 @@ impl Mempool {
         //
         // The additional bookkeeping can be implemented once we have fee related strategies. KISS.
 
-        let mut selected = ProposedBatchNode::default();
+        let mut selected = SelectedBatch::builder();
         let mut budget = self.config.batch_budget;
 
         let mut candidates = self.nodes.txs.values();
 
         'next: while let Some(candidate) = candidates.next() {
-            if selected.contains(candidate.id()) {
+            if selected.contains(&candidate.id()) {
                 continue 'next;
             }
 
@@ -294,7 +295,9 @@ impl Mempool {
                 match parent {
                     // TODO(mirko): Once user batches are supported, they will also need to be
                     // checked here.
-                    NodeId::Transaction(parent) if !selected.contains(parent) => continue 'next,
+                    NodeId::Transaction(parent) if !selected.contains(&parent) => {
+                        continue 'next;
+                    },
                     NodeId::Transaction(_)
                     | NodeId::ProposedBatch(_)
                     | NodeId::ProvenBatch(_)
@@ -313,11 +316,12 @@ impl Mempool {
         if selected.is_empty() {
             return None;
         }
+        let selected = selected.build();
 
-        let batch_id = selected.calculate_id();
-        let batch_txs = selected.transactions().cloned().collect::<Vec<_>>();
+        let batch = ProposedBatchNode::new(selected.clone());
+        let batch_id = batch.batch_id();
 
-        for tx in &batch_txs {
+        for tx in batch.transactions() {
             let node =
                 self.nodes.txs.remove(&tx.id()).expect("selected transaction node must exist");
             self.state.remove(&node);
@@ -327,13 +331,13 @@ impl Mempool {
                 "Transaction selected for inclusion in batch"
             );
         }
-        self.state.insert(NodeId::ProposedBatch(batch_id), &selected);
-        self.nodes.proposed_batches.insert(batch_id, selected);
+        self.state.insert(NodeId::ProposedBatch(batch_id), &batch);
+        self.nodes.proposed_batches.insert(batch_id, batch);
 
         // TODO(mirko): Selecting a batch can unblock user batches, which should be checked here.
 
         self.inject_telemetry();
-        Some((batch_id, batch_txs))
+        Some(selected)
     }
 
     /// Drops the proposed batch and all of its descendants.
@@ -415,7 +419,8 @@ impl Mempool {
             self.nodes.proposed_block.as_ref().unwrap().0
         );
 
-        let mut selected = BlockNode::default();
+        let block_number = self.chain_tip.child();
+        let mut selected = BlockNode::new(block_number);
         let mut budget = self.config.block_budget;
         let mut candidates = self.nodes.proven_batches.values();
 
@@ -443,7 +448,6 @@ impl Mempool {
             selected.push(candidate.clone());
         }
 
-        let block_number = self.chain_tip.child();
         // Replace the batches with the block in state and nodes.
         for batch in selected.batches() {
             // SAFETY: Selected batches came from nodes, and are unique.